Initialize the main page with a first content and a global CSS
