+++ public/page-data/404/page-data.json
@@ -1,0 +1,1 @@
+{"componentChunkName":"component---src-pages-404-js","path":"/404/","result":{"pageContext":{}},"staticQueryHashes":[]}
