+++ src/styles/global.css
@@ -1,0 +1,13 @@
+main {
+    margin: auto;
+    max-width: 900px;
+    font-family: sans-serif;
+    color: #232129;
+    padding: 96px;
+}
+h1 {
+    color: rebeccapurple;
+}
+p {
+    marginBottom: 48,Ò
+}
